Add nodejs starter files
